requirements_aks.txt
azureml-core==1.41.0
azureml-pipeline==1.41.0
argcomplete==1.12.3
azure-core==1.24.0
azure-identity==1.7.0
azure-mgmt-storage==16.0.0
azure-keyvault==1.1.0
azure-keyvault-secrets==4.4.0
azure-storage==0.36.0
azure-storage-blob==12.11.0
azure-storage-common==1.4.2
numpy==1.23.3
pandas==1.4.2
pyodbc==4.0.35
PyYAML==6.0
requests==2.27.1
requests-oauthlib==1.3.1